+++ bootstrap/src/sun/nio/ch/lincheck/EventTracker.java
@@ -32,10 +32,7 @@ public interface EventTracker {
     void beforeNewObjectCreation(String className);
     void afterNewObjectCreation(Object obj);
 
-    void updateSnapshotOnFieldAccess(Object obj, String className, String fieldName, int codeLocation);
-    void updateSnapshotOnArrayElementAccess(Object array, int index, int codeLocation);
     void updateSnapshotWithEnergeticTracking(Object[] objs);
-
     boolean beforeReadField(Object obj, String className, String fieldName, int codeLocation,
                             boolean isStatic, boolean isFinal);
     boolean beforeReadArrayElement(Object array, int index, int codeLocation);
