+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/CTestConfiguration.kt
@@ -95,7 +95,6 @@ internal fun createFromTestClassAnnotations(testClass: Class<*>): List<CTestConf
                     checkObstructionFreedom = ann.checkObstructionFreedom,
                     hangingDetectionThreshold = ann.hangingDetectionThreshold,
                     invocationsPerIteration = ann.invocationsPerIteration,
-                    restoreStaticMemory = ann.restoreStaticMemory,
                     guarantees = ManagedCTestConfiguration.DEFAULT_GUARANTEES,
                     minimizeFailedScenario = ann.minimizeFailedScenario,
                     sequentialSpecification = chooseSequentialSpecification(
