+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/strategy/managed/ManagedStrategy.kt
@@ -219,11 +219,11 @@ abstract class ManagedStrategy(
         parkingTracker.reset()
     }
 
-    override fun restoreStaticMemorySnapshot() {
-        if (testCfg.restoreStaticMemory) {
-            staticMemorySnapshot.restoreValues()
-            super.restoreStaticMemorySnapshot()
-        }
+    /**
+     * Restores recorded values of all memory reachable from static state.
+     */
+    fun restoreStaticMemorySnapshot() {
+        staticMemorySnapshot.restoreValues()
     }
 
     /**
@@ -232,8 +232,13 @@ abstract class ManagedStrategy(
     override fun runInvocation(): InvocationResult {
         while (true) {
             initializeInvocation()
-            restoreStaticMemorySnapshot()
-            val result = runner.run()
+            val result: InvocationResult
+            try {
+                result = runner.run()
+            }
+            finally {
+                restoreStaticMemorySnapshot()
+            }
             // In case the runner detects a deadlock, some threads can still manipulate the current strategy,
             // so we're not interested in suddenInvocationResult in this case
             // and immediately return RunnerTimeoutInvocationResult.
@@ -753,6 +758,7 @@ abstract class ManagedStrategy(
      */
     override fun beforeReadField(obj: Any?, className: String, fieldName: String, codeLocation: Int,
                                  isStatic: Boolean, isFinal: Boolean) = runInIgnoredSection {
+         updateSnapshotOnFieldAccess(obj, className.canonicalClassName, fieldName)
         // We need to ensure all the classes related to the reading object are instrumented.
         // The following call checks all the static fields.
         if (isStatic) {
@@ -789,6 +795,7 @@ abstract class ManagedStrategy(
 
     /** Returns <code>true</code> if a switch point is created. */
     override fun beforeReadArrayElement(array: Any, index: Int, codeLocation: Int): Boolean = runInIgnoredSection {
+        updateSnapshotOnArrayElementAccess(array, index)
         if (!objectTracker.shouldTrackObjectAccess(array)) {
             return@runInIgnoredSection false
         }
@@ -824,6 +831,7 @@ abstract class ManagedStrategy(
 
     override fun beforeWriteField(obj: Any?, className: String, fieldName: String, value: Any?, codeLocation: Int,
                                   isStatic: Boolean, isFinal: Boolean): Boolean = runInIgnoredSection {
+        updateSnapshotOnFieldAccess(obj, className.canonicalClassName, fieldName)
         objectTracker.registerObjectLink(fromObject = obj ?: StaticObject, toObject = value)
         if (!objectTracker.shouldTrackObjectAccess(obj ?: StaticObject)) {
             return@runInIgnoredSection false
@@ -853,6 +861,7 @@ abstract class ManagedStrategy(
     }
 
     override fun beforeWriteArrayElement(array: Any, index: Int, value: Any?, codeLocation: Int): Boolean = runInIgnoredSection {
+        updateSnapshotOnArrayElementAccess(array, index)
         objectTracker.registerObjectLink(fromObject = array, toObject = value)
         if (!objectTracker.shouldTrackObjectAccess(array)) {
             return@runInIgnoredSection false
@@ -924,19 +933,15 @@ abstract class ManagedStrategy(
      * Tracks a specific field of an [obj], if the [obj] is either `null` (which means that field is static),
      * or one this objects which contains it is already stored.
      */
-    override fun updateSnapshotOnFieldAccess(obj: Any?, className: String, fieldName: String, codeLocation: Int) = runInIgnoredSection {
-        if (testCfg.restoreStaticMemory) {
-            staticMemorySnapshot.trackField(obj, className, fieldName)
-        }
+    fun updateSnapshotOnFieldAccess(obj: Any?, className: String, fieldName: String) = runInIgnoredSection {
+        staticMemorySnapshot.trackField(obj, className, fieldName)
     }
 
     /**
      * Tracks a specific [array] element at [index], if the [array] is already tracked.
      */
-    override fun updateSnapshotOnArrayElementAccess(array: Any, index: Int, codeLocation: Int) = runInIgnoredSection {
-        if (testCfg.restoreStaticMemory) {
-            staticMemorySnapshot.trackArrayCell(array, index)
-        }
+    fun updateSnapshotOnArrayElementAccess(array: Any, index: Int) = runInIgnoredSection {
+        staticMemorySnapshot.trackArrayCell(array, index)
     }
 
     /**
@@ -944,9 +949,7 @@ abstract class ManagedStrategy(
      * Required as a trick to overcome issue with leaking this in constructors, see https://github.com/JetBrains/lincheck/issues/424.
      */
     override fun updateSnapshotWithEnergeticTracking(objs: Array<Any?>) = runInIgnoredSection {
-        if (testCfg.restoreStaticMemory) {
-            staticMemorySnapshot.trackObjects(objs)
-        }
+        staticMemorySnapshot.trackObjects(objs)
     }
 
     private fun methodGuaranteeType(owner: Any?, className: String, methodName: String): ManagedGuaranteeType? = runInIgnoredSection {
